+++ main/src/arg-group-factories/__tests__/cli-string-choice-arg-group.test.ts
@@ -8,7 +8,7 @@ const placeholder = '<special>';
 const required = false;
 
 const argGroup = CliStringChoiceArgGroup({
-	choices: ['foo', 'bar'],
+	choices: ['foo', 'bar'] as const,
 	description,
 	hidden,
 	placeholder,